Relax test
The exact number of lines in the mmCIF output file
will depend on the exact versions of IHM, IMP
and PMI that we use, so just ensure a reasonably
large file was created for now. Ultimately we will
employ IHM itself (once its read support is complete)
to read the file back in and check its contents.